final cactus step

+++ cactus_pipeline.sh
@@ -13,7 +13,7 @@ source /scratch/sdubois/Stage_M2/cactus-bin-v2.4.0/cactus_env/bin/activate
 # $4 -> a temp name
 
 # destroy any .jobstore
-rm -r ./jobstore_$4
+[ -d ./jobstore_$4 ] && rm -r ./jobstore_$4
 
 # creating GFA1 file (SV graph)
 cactus-minigraph ./jobstore_$4 $1 $2.gfa --reference $3
@@ -22,7 +22,10 @@ cactus-minigraph ./jobstore_$4 $1 $2.gfa --reference $3
 cactus-graphmap ./jobstore_$4 $1 $2.gfa $2.paf  --reference $3 --outputFasta $2.sv.gfa.fa.gz
 
 # creating hal (cactus base alignment)
-cactus-align ./jobstore_$4 $1 $2.paf $2.hal --pangenome --outGFA --reference $3 
+cactus-align ./jobstore_$4 $1 $2.paf $2.hal --pangenome --outGFA --outVG --reference $3 
+
+# final step
+cactus-graphmap-join ./jobstore_$4 --gfa --vg --outDir ./$2 --outName $4 --reference $3 --vcf --giraffe
 
 # $1 must be a txt file from the format 
 # Diploid sample:
